# perf: remove empty HashMap instances from TrieUpdates and HashedPostState